% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_weights.R
\name{compute_weights_BMA}
\alias{compute_weights_BMA}
\title{Compute the model evidence/marginal likelihood via bridge sampling (or via the harmonic mean estimator if bridge sampling fails)}
\usage{
compute_weights_BMA(stan_fits, mortality_models)
}
\arguments{
\item{stan_fits}{list of Stan model fits where the marginal likelihood was computed via bridge sampling}

\item{mortality_models}{vector of mortality models names}
}
\value{
data frame with model evidence for BMA
}
\description{
Compute the model evidence/marginal likelihood via bridge sampling (or via the harmonic mean estimator if bridge sampling fails)
}

