% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mortality_model.R
\name{extract_map}
\alias{extract_map}
\title{Function to get the a posterior means of the parameters based on a stanfit object}
\usage{
extract_map(stan_fit)
}
\arguments{
\item{stan_fit}{a stanfit object}
}
\value{
named list with the point estimates of the parameters
}
\description{
Function to get the a posterior means of the parameters based on a stanfit object
}
