% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRMaleData.R
\docType{data}
\name{FRMaleData}
\alias{FRMaleData}
\title{Deaths and Exposures Data of French Males}
\format{
A list of 2 matrices
\describe{
\item{Dxt}{Matrix of deaths, with 111 rows (ages) and 202 columns (years)}
\item{Ext}{Matrix of exposures, with 111 rows (ages) and 202 columns (years)}
}
}
\source{
Human Mortality Database \url{https://www.mortality.org/}
}
\usage{
FRMaleData
}
\description{
List containing deaths and exposures of French Male for ages 0-110 and years 1816-2017.
}
\references{
Human Mortality Database (2011).   University of California,  Berkeley (USA), and Max Planck Institute
for Demographic Research (Germany). Available at \url{https://www.mortality.org/}
}
\keyword{datasets}
