% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6_stan.R
\name{m6_stan}
\alias{m6_stan}
\title{Bayesian M6 model with Stan}
\usage{
m6_stan(
  death,
  exposure,
  forecast,
  age,
  validation = 0,
  family = c("poisson", "nb"),
  ...
)
}
\arguments{
\item{death}{Matrix of deaths}

\item{exposure}{Matrix of exposures}

\item{forecast}{Number of years to forecast}

\item{age}{Vector of ages}

\item{validation}{Number of years for validation}

\item{family}{specifies the random component of the mortality model. \code{"Poisson"} assumes a
Poisson model with log link and \code{"nb"} assumes a negative-binomial model
with log link and overdispersion parameter}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian M6 model with Stan
}
