% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_weights.R
\name{mortality_weights}
\alias{mortality_weights}
\title{Mortality Model averaging via stacking of predictive distributions, pseudo-BMA weighting and pseudo-BMA+
weighting with the Bayesian bootstrap. Based on Yao et al. (2018) but adapted by Barigou et al. (2020)
for mortality forecasting.}
\usage{
mortality_weights(X)
}
\arguments{
\item{X}{A list of stanfit objects}
}
\value{
A matrix containing one weight for each model and each approach.
}
\description{
Mortality Model averaging via stacking of predictive distributions, pseudo-BMA weighting and pseudo-BMA+
weighting with the Bayesian bootstrap. Based on Yao et al. (2018) but adapted by Barigou et al. (2020)
for mortality forecasting.
}
