% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mortality_data.R
\name{sim_death_apc}
\alias{sim_death_apc}
\title{sim_death_apc simulate death counts from Age-Period-Cohort mortality model}
\usage{
sim_death_apc(a, k, g, phi, years, ages, exposure)
}
\arguments{
\item{a}{vector of age component}

\item{k}{vector of period component}

\item{g}{vector of cohort component}

\item{phi}{dispersion parameter}

\item{years}{vector of calendar years}

\item{ages}{vectors of ages}

\item{exposure}{matrix of exposure data}
}
\value{
matrix of death count
}
\description{
sim_death_apc simulate death counts from Age-Period-Cohort mortality model
}
