% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mortality_data.R
\name{sim_death_cbd}
\alias{sim_death_cbd}
\title{Simulation of death counts from the CBD model}
\usage{
sim_death_cbd(k, k2, phi, years, ages, exposure)
}
\arguments{
\item{k}{first vector of period component}

\item{k2}{second vector of period component}

\item{phi}{dispersion parameter}

\item{years}{vector of calendar years}

\item{ages}{vectors of ages}

\item{exposure}{matrix of exposure data}
}
\value{
matrix of death count
}
\description{
Simulation of death counts from the CBD model
}
