% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mortality_data.R
\name{sim_death_lc}
\alias{sim_death_lc}
\title{sim_death_lc simulate death counts from Lee-Carter mortality model}
\usage{
sim_death_lc(a, b, k, phi, exposure)
}
\arguments{
\item{a}{vector of age component}

\item{b}{vector of age/year component}

\item{k}{vector of year component}

\item{phi}{dispersion parameter}

\item{exposure}{matrix of exposure data}
}
\value{
matrix of death count
}
\description{
sim_death_lc simulate death counts from Lee-Carter mortality model
}
