% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mortality_data.R
\name{sim_death_m6}
\alias{sim_death_m6}
\title{sim_death_m6 simulates death counts from the M6 model}
\usage{
sim_death_m6(k, k2, g, phi, years, ages, exposure)
}
\arguments{
\item{k}{first vector of period component}

\item{k2}{second vector of period component}

\item{g}{vector of cohort component}

\item{phi}{dispersion parameter}

\item{years}{vector of calendar years}

\item{ages}{vectors of ages}

\item{exposure}{matrix of exposure data}
}
\value{
matrix of death count
}
\description{
sim_death_m6 simulates death counts from the M6 model
}
