% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mortality_data.R
\name{sim_mortality_data}
\alias{sim_mortality_data}
\title{sim_mortality_data simulate mortality data from various models}
\usage{
sim_mortality_data(a, k, b, g, phi, years, ages, exposure, mortality_model)
}
\arguments{
\item{a}{vector of age component}

\item{k}{vector of time component}

\item{b}{vector of age/time component}

\item{g}{vector of cohort component}

\item{phi}{dispersioon parameter}

\item{years}{vector of calendar year}

\item{ages}{vector of ages}

\item{exposure}{matrix of exposure}

\item{mortality_model}{name of the mortality model that we simulate from}
}
\value{

}
\description{
sim_mortality_data simulate mortality data from various models
}
