% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mortality_model.R
\name{extract_map}
\alias{extract_map}
\title{extract_map: function to get the mean a posteriori of the parameters based on a stanfit object}
\usage{
extract_map(stan_fit)
}
\arguments{
\item{stan_fit}{a stanfit object}
}
\value{
named list with the point estimates of the parameters
}
\description{
extract_map: function to get the mean a posteriori of the parameters based on a stanfit object
}
