% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mo_mo.R
\name{fit_mo_mo}
\alias{fit_mo_mo}
\title{fit_mo_mo return the posterior samples as stan object of the fitted mortality model}
\usage{
fit_mo_mo(
  mortality_model = "lc",
  death = deathGBR,
  exposure = exposureGBR,
  ages = 50:90,
  validation = 0,
  forecast = 1,
  family = "nb",
  chains = 1,
  cores = 4
)
}
\arguments{
\item{mortality_model}{name of teh mortality model}

\item{death}{death matrix}

\item{exposure}{exposure matrix}

\item{ages}{vector of ages}

\item{validation}{size of the validation set}

\item{forecast}{number of calendar year to be forecast}

\item{family}{underlying count distribution}

\item{chains}{number of Markov chains}

\item{cores}{number of cores used}
}
\value{
a stanfit object
}
\description{
fit_mo_mo return the posterior samples as stan object of the fitted mortality model
}
