% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_plot.R
\name{forecasting_plot}
\alias{forecasting_plot}
\title{Fanplot for the mortality predictions}
\usage{
forecasting_plot(stan_fit, ages, years, death, exposure, ages.plot)
}
\arguments{
\item{stan_fit}{stanfit object}

\item{ages}{range of ages}

\item{years}{range of years}

\item{death}{matrix of observed deaths}

\item{exposure}{matrix of observed exposures}

\item{ages.plot}{ages to be plotted}
}
\description{
The function generates a fan plot representing the forecast death rates based on the \code{ggplot2} and
\code{ggfan} R packages.
}
\examples{

\dontrun{
years <- 1980:2018
ages <- 50:90
death <- load_HMD_data('BEL', 'Deaths_1x1', years, ages, "Male")$mat
exposure <- load_HMD_data('BEL', 'Exposures_1x1', years, ages, "Male")$mat
fitLC<-lc_stan(death = deathFR,exposure=exposureFR, forecast = 10, family = "poisson",chains=1,cores=1)
forecasting_plot(fitLC,ages,years,death,exposure,c(65,75,85))
}

}
