% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_HMD_data.R
\name{load_HMD_data}
\alias{load_HMD_data}
\title{load_HMD_data: a function to download mortality data from the HMD website}
\usage{
load_HMD_data(CNTRY, item, cal_year_range, age_range, gender)
}
\arguments{
\item{CNTRY}{String, HMD country name for instance 'FRATNP'}

\item{item}{String, HMD data type for instance 'Deaths_1x1'}

\item{cal_year_range}{Array, calendar year of interest, for instance
1970:2017}

\item{age_range}{Array, ages of interest, for instance 50:90}

\item{gender}{String, gender, for instance "Male"}
}
\value{
a named list containing the data either under a datafreame, a matrix
or a vector
}
\description{
load_HMD_data: a function to download mortality data from the HMD website
}
\examples{
death_fra <- retrieve_HMD_data(CNTRY, 'Deaths_1x1', 1970:2017, 50:90, "Male")
exposure_fra <- retrieve_HMD_data(CNTRY, 'Exposures_1x1', 1970:2017, 50:90, "Male")
}
