% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mortality_data.R
\name{sim_death_mix_cbd_rh}
\alias{sim_death_mix_cbd_rh}
\title{sim_death_mix_cbd_rh simulates death counts from a hybrid model that average the mortality
rates from cbd and rh model}
\usage{
sim_death_mix_cbd_rh(params_cbd, params_rh, years, ages, exposure, q)
}
\arguments{
\item{params_cbd}{named lsit that contains the parameters of the cbd model}

\item{params_rh}{named lsit that contains the parameters of the rh model}

\item{years}{vector of calendar year}

\item{ages}{vector of ages}

\item{exposure}{matrix of exposure data}

\item{q}{mixing parameter (0 <- rh, 1 <- cbd)}
}
\value{
matrix of death count
}
\description{
sim_death_mix_cbd_rh simulates death counts from a hybrid model that average the mortality
rates from cbd and rh model
}
