% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mortality_data.R
\name{sim_mortality_data}
\alias{sim_mortality_data}
\title{Simulation of mortality data from various models}
\usage{
sim_mortality_data(a, k, k2, b, g, phi, years, ages, exposure, mortality_model)
}
\arguments{
\item{a}{vector of age component}

\item{k}{first vector of time component}

\item{k2}{second vector of time component}

\item{b}{vector of age/time component}

\item{g}{vector of cohort component}

\item{phi}{dispersion parameter}

\item{years}{vector of calendar year}

\item{ages}{vector of ages}

\item{exposure}{matrix of exposure}

\item{mortality_model}{name of the mortality model that we simulate from}
}
\value{
matrix of death counts
}
\description{
Simulation of mortality data from various models
}
