% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanMoMo-package.R
\docType{package}
\name{StanMoMo-package}
\alias{StanMoMo-package}
\alias{StanMoMo}
\title{The 'StanMoMo' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
